refact: util node/registry split
