+++ crates/pm/src/model/mod.rs
@@ -1,3 +1,5 @@
+pub mod node;
+pub mod override_rule;
 pub mod package;
 pub mod package_lock;
 pub mod package_manifest;
