+++ crates/pm/src/service/mod.rs
@@ -1,9 +1,12 @@
 pub mod binary;
+pub mod cache;
 pub mod config;
 pub mod dependency_graph;
 pub mod execute;
+pub mod http_client;
 pub mod install;
 pub mod package;
+pub mod registry;
 pub mod script;
 pub mod update;
 pub mod workspace;
