+++ crates/pm/src/util/mod.rs
@@ -3,10 +3,11 @@ pub mod cache;
 pub mod cloner;
 pub mod config;
 pub mod downloader;
+pub mod format_print;
 pub mod json;
 pub mod linker;
 pub mod logger;
-pub mod node;
+pub mod node_search;
 pub mod package_installer;
 pub mod registry;
 pub mod relative_path;
@@ -14,5 +15,3 @@ pub mod retry;
 pub mod save_type;
 pub mod semver;
 pub mod timer;
-
-pub mod format_print;
